requirements.txt
pyserial>=3